suppress warning in test build when used as a library
My rustc (1.36) complains about a redundant `#[macro_use]` directive for
serde_derive in lib.rs when I run `cargo test` on a project that uses ipmpsc
and also uses serde_derive itself.  However that directive is required when
the ipmpsc tests are run directly.

There may be a better way to handle this, but explictly `allow`ing the warning
does the job.