switch from `failure` to `thiserror`/`anyhow`
The `failure` crate has been deprecated, and the way we were using it was
inappropriate for a library anyway.  Now we use `thiserror` to derive `Error`
and `Debug` for a custom error enum, plus `anyhow` for the tests.  This allows
apps to `match` on errors rather than the downcast if/else chains required
previously.

Version bumped to 0.3.0 since this is a breaking API change.