fix race condition in Sender::send_0
We need to reload the write offset each time we loop in case another writer
updates it while we were waiting.

I've updated the tests to cover any number of simultaneous senders from 1 to 5.
Previously, we only tested the single sender case, which is why the race
condition wasn't apparent.