add `flags` field to `Header`
For now, this just allows us to label the ring buffer as having been created and
initialized by a 32-bit or 64-bit process, ensuring that any process with a
different word size will fail immediately with a clear diagnostic if it tries to
open that ring buffer.

In the future, we can add other things to that field, e.g. to indicate the ring
buffer was created by an incompatible version of `ipmpsc`.